[Driver][SYCL] Allow for -o output with dependency generation
Previous fix to for dependency generation with -fsycl caused a problem
when emitting an output file and dependency info.  This was causing a
multiple output error, due to the extra preprocess step which is counted
as a compilation step.  Allow for this new step to not count as an
official output step.