[SYCL][NFC] Drop on-device tests in favor of llvm-test-suite (#4393)
The original motivation for introducing on-device tests was two-fold:

- Provide a way to test product changes together with E2E test changes
for features in active development, these tests would then be moved to
llvm-test-suite once the feature is considered stable. Frequently,
instead of using on-device tests (or in case an existing llvm-test-suite
test has to be modified) developers would create a pull request in each
repo, link them via a comment, then they would be either merged
simultaneously (while relying on local testing and intel/llvm post-commit
results for the modified/added tests) or one after the other.
- As a way to have some sanity E2E checks in intel/llvm. On-device
tests haven't been used in this manner, we've been relying on component
tests instead.

This patch drops in-tree on-device tests in favor of llvm-test-suite
in order to reduce in-tree test running time.