Fix device-mangling to be more consistent, other review comments
The Unique-stable-** now make sure to use 'device mangling number'
always all of the time, rather than counting on Itanium to always set it
to 0, which would result in us using the 'lambda number'.  This patch
instead changes our ItaniumCXXABI to always set the device mangling
number to match the lambda mangling number, so we can always use the
device mangling number if it exists.

This patch also fixes all of the review comments so far.