[SYCL] Fix several cases of locking the same object twice
* piEventsWait is called during piEnqueueMemMap/piEnqueueMemUnmap which
clauses same queue being locked twice. Introduce EventsWait helper
function to check if queue was already locked.

* piEnqueueKernelLaunch function needs both queue and kernel object to
be locked. But getAvailableCommandList may result in resetting
commandlist which contains the same kernel, resetting starts event
cleanup and release of the kernel associated with event. piKernelRelease
may lock same kernel which was locked at the start of
piEnqueueKernelLaunch.
Since it is enough to lock a queue for getAvailableCommandList then move
a call to this function to the beginning of piEnqueueKernelLaunch and
don't lock a kernel before call.