Actual fix
It's not possible to provide a template specialization for
is_device_copyable with kernel lambda since lambda types don't
have the name. To fix this issue, we create a functor class
RoundedRangeKernel and provide a is_device_copyable trait for this
class, which simply forwards check to nested kernel lambda.