[SYCL][Driver] Fix autodetection of the -Xsycl-target-backend triple (#4463)
https://github.com/intel/llvm/pull/4175 introduced automatic addition
of the generic spir64 device target when any section of the input
objects had this triple assigned to it. As a result, the actual list
of toolchains started exceeding the user-provided one by 1 item.

After https://github.com/intel/llvm/pull/4239, the above became a
problem. The dispatch of -Xsycl-target-* arguments started happening
earlier in theflow, which broke the following use-case:
```
clang++ -fsycl -fsycl-targets=spir64_gen gen-obj.o
gen-and-spir64-obj.o -Xsycl-target-backend "-device *"
```
A fix for now is to ignore the autodetected spir64 target when
propagating the -Xsycl-target-backend arguments. A permanent solution
would involve a re-design of -Xsycl-target-backend handling so that it
took place only once in the flow, or belating the addition of the
autodetected generic triple into the list of device targets.

Signed-off-by: Artem Gindinson <artem.gindinson@intel.com>