[SYCL][Driver] Diagnose incorrect inputs to CLI options
Instead of assertions and silent fallthrough scenarios, incorrect
values to SYCL options should lead to user-friendly diagnostics.

The long-term fix for this should be a SYCL-agnostic solution in the
LLVM comunity code - allowed values to driver options are known at the
TableGen level via the `Values<>` class, so the validation should be
happening there too. This would minimize code duplication for
individual compilation pipelines and bring the TableGen CLI library
closer to the `llvm::cl` functionality.

Signed-off-by: Artem Gindinson <artem.gindinson@intel.com>