[SYCL] Add Level-Zero interop with specification of ownership for kernel_bundle.
The L0 backend interop functions take the ownership of the passed L0 handles.
This is problematic for some cases as the users can not use the original handle
when they go out of scope and destroyed. This PR provides a way for users of
SYCL - L0 interop to retain the ownership of the L0 handles for the kernel
and kernel_bundle.

Signed-off-by: rehana begam <rehana.begam@intel.com>