[SYCL] Align id and range class implementation with SYCL specification (#4538)
The DPC++ "range" implementation has a conversion to "id",
which is not in the spec, and "id" implementation has a
conversion to "range", which is also not in the spec.

As for operators:
The 16 binary operators "+, -, *, /, %, <<, >>, &, |, ^, &&, ||, <, >, <=, >=" 
are defined as either member functions or hidden friend
functions depending on the arguments. However, the spec
defines these operators as hidden friend functions for all
argument variations.
The 16 binary operators are defined as templates that take
any integral type, but the spec defines these operators only
for "size_t". The 10 assignment operators
"+=, -=, *=, /=, %=, <<=, >>=, &=, |=, ^=" are defined as
member functions, but the spec defines them as hidden friend
functions.

In addition, DPC++ is missing the following hidden friend
functions which are defined in the spec:
Unary "+" and "-".
Prefix "++" and "--"
Postfix "++" and "--"