Revert "Add validation of vloada_half instruction (#1138)"
This reverts commit 1a7f1420db4dc21a9628b85e9c06073c718ebfc4.

The vloada_halfn functions return float vectors:

   floatn vloada_halfn(size_t offset, const __global half *p)
   floatn vloada_halfn(size_t offset, const __local half *p)
   floatn vloada_halfn(size_t offset, const __constant half *p)
   floatn vloada_halfn(size_t offset, const __private half *p)

   For the generic address space:
   floatn vloada_halfn(size_t offset, const half *p)

These functions exist specifically for implementations that lack the
cl_khr_fp16 extension, and so do not include the halfn vector types,
but do include a scalar half type for use as a pointer argument *only*.

Likewise, in the SPIR-V OpenCL Extended Instruction Set, the result type
for vloada_halfn is defined as follows:

   Result Type must be vector(2,3,4,8,16) of float values.

The assertion added to isBuiltinTransToExtInst(...) is therefore
incorrect.

The vloada_halfn SPIR-V OpenCL Extended Instruction Set instructions
that are emitted for calls to these builtins are accepted by the
SPIR-V Validator.

Original commit:
https://github.com/KhronosGroup/SPIRV-LLVM-Translator/commit/2b95385