[SYCL][libclc] Fix incorrectly mangled names (#4600)
Fixes #4500. There are some manually mangled names for querying image sizes in libclc that are incorrect. Information on return types from mangled name does not match the actual return type. In one case the return type also seem to be wrong - this one is needed in the failed CTS test.