[SYCL] Fix zero dimension accessors on FPGA in AOT mode (#4458)
Zero dimension accessors aren't working on FPGA with atomic data.
When aot compiling for FPGA we use address space
global_device_space instead of just global_space. This seems to be
confusing the specialization for the zero dimension accessor. Not
overspecifying the address space fixes the problem with no other
change in functionality.

Signed-off-by: Chris Perkins <chris.perkins@intel.com>