[SYCL] Throw exception when kernel is missing from kernel bundle
According to the SYCL 2020 specification, if a kernel bundle is set to
be used by a command group, as set through use_kernel_bundle of the
associated handler, then submitting a kernel that does not exist in any
suitable device image of the kernel bundle should throw a synchronous
sycl::exception with error code errc::kernel_not_supported.

These changes makes the kernel invocation command throw a
sycl::exception with error code errc::kernel_not_supported if the
kernel bundle in use has no suitable device images with the specified
kernel.

Additionally these changes remove the undefined behavior and assertion
for kernel_bundle::begin and kernel_bundle::end when the vector of
device images is empty.

Signed-off-by: Steffen Larsen <steffen.larsen@intel.com>