[Driver][SYCL] Fix offload-bundler and offload-deps triples
The patch 98380762c3b734c23d206182605ab9e035c93caa changed triple
processing so gpu arch could be deducted even without extra `-` so we no
longer need to add padding, it also seems like SYCL was inadvertently
removed from the branch adding in the bound arch to the triple.

So this patch fixes adding the bound arch in the triple when using SYCL,
removes leftover triple padding code in the offload deps command.

The test was also updated accordingly and it now also checks the triple
used for `clang-offload-deps` so that we can hopefully catch mismatch
between the two earlier in the future.

This fixes #4307