[SYCL] Prevent stream buffer leak on constructor exception (#4594)
The `sycl::stream` class is currently throwing a `sycl::invalid_parameter_error` exception in the constructor when the `MaxStatementSize` parameter exceeds the `MAX_STATEMENT_SIZE` limitation. However, this exception is thrown after
the underlying `stream_impl` object and the corresponding accessors are initialized. `stream_impl` allocates the buffers for the stream but leaves it to the scheduler to deallocate it. Since the exception is thrown prior to registering the stream with the command-group handler, the allocated buffers will leak if the exception is thrown.

These changes prevents the memory leak by making the check prior to the initialization of the `stream_impl`, which in turn avoids the creation of the stream buffers and accessors if the check fails. This also avoids allocating memory that will never be used.