[SYCL] Detach allocas from their dependencies regardless of linked alloca presence (#4573)
cleanupCommandsForRecord() removed allocation commands from the users of their
dependencies only when there was a linked alloca command present. This has caused
dangling pointers in the graph which resulted in invalid reads hence sporadic
segfaults. This PR detaches alloca commands from their dependencies regardless of
whether a linked alloca is present or not.