Fix mistranslation of OpAtomicCompareExchangeWeak
The semantics of OpAtomicCompareExchangeWeak are the same as those of
OpAtomicCompareExchange.  Each of the ops should be translated to
atomic_compare_exchange_strong_explicit() when using OpenCL 2.0 builtins,
and not atomic_compare_exchange_weak_explicit(), as use of the latter may
result in unreported spurious failures which are not detectable when
using OpAtomicCompareExchangeWeak, which does not give a direct
indication of its success.

Original commit:
https://github.com/KhronosGroup/SPIRV-LLVM-Translator/commit/b226eb5