[SYCL] Implement SYCL2020 target enumeration: replace global_buffer with device (#4588)
The SYCL 2020 spec deprecates four of the enumeration
values and replaces them with two new ones. While mostly
a fairly straightforward change, I realize this is beginning to
touch a lot of code, which will make it difficult to review. So
I'm breaking it into smaller chunks.
We've already added device::target in a previous PR. In this
PR we are simply marking target::global_buffer as deprecated
and updating its uses in the code base.

Signed-off-by: Chris Perkins <chris.perkins@intel.com>