[SYCL] Fix an error on host when big image is used on opencl:gpu
An exception was thrown on host because during the host allocation
the image seemed too big for host even though the target device
could process such big image. Because the limits to image height,
width,depth are arbitrary on HOST, they have been relaxed to MAX_INT.

The existing check for image sizes is still sufficient when image
is passed to device because the image parameters are verified twice:
one time for host and another time for device.

Signed-off-by: Vyacheslav N Klochkov <vyacheslav.n.klochkov@intel.com>