Don't emit UserSemantic with full annotation str if FPGA decorations found
https://github.com/KhronosGroup/SPIRV-LLVM-Translator/pull/1206 with a
change of 'AnnotatedDecoration' to 'AnnotatedCode' brought a regression,
when the annotation information started to be duplicated in both
UserSemantic decoration and FPGA-specific decorations. This patch fixes
that.

Yet if there is an annotation string that is split in blocks like Intel
FPGA annotation, it's not necessarily an FPGA annotation. Translate the
whole string as UserSemantic decoration in this case.

Signed-off-by: Dmitry Sidorov <dmitry.sidorov@intel.com>

Original commit:
https://github.com/KhronosGroup/SPIRV-LLVM-Translator/commit/336248f