[SYCL][HIP] Add HIP arch argument for lit tests (#4676)
When using the HIP backend for an AMD platform the gpu's architecture must be sepcified with `-Xsycl-target-backend=amdgcn-amd-amdhsa --offload-arch=<target>`. This patch adds an option to `buildbot/configure.py` which allows the user to state the gpu architecture for lit tests when running `check-sycl-hip`. 
Relevant messages and docs are updated within the patch.

This patch is needed even if the tests do not run on hardware, as the amd architecture must be specified for compilation.