[SYCL] Align device::create_sub_devices exceptions with SYCL 2020 (#4670)
According to the SYCL 2020 spec device::create_sub_devices
should throw an exception with either
errc::feature_not_supported or errc::invalid error codes. The
feature_not_supported side is working fine. But for invalid
values, we presently just lean on the default exception thrown
by plugin::call() which is always a runtime_error exception
subclass. Here we are adding new specializations for
plugin::call() and plugin::checkPiResult that accept sycl::errc
values and use those specializations from
device::create_sub_devices() to ensure that errc::invalid is the
error code in the exception (instead of errc::runtime)

This bug is presently tripping the SYCL-CTS tests, thus the need to fix it.

Signed-off-by: Chris Perkins <chris.perkins@intel.com>