[SYCL] Fix sycl-ls device indexing
This fixes two issues, first `backend::all` is not guaranteed to be the
last entry in the backend types enum and shouldn't be used as such. This
caused issues for newer backends such as HIP (see #4698).

The second issue is that the `DeviceNums` vector wasn't being reset
between the short output and the verbose output which meant that the
device indices printed in the verbose output were incorrect.