[SYCL] Tie BoundArch to Toolchain  (#4673)
These changes implement SYCLTargetInfoList which stores both the toolchain and the bound arch. This prevents BoundArchs being incorrectly assigned when actions and dependencies are created. This sets up the possibility of having multiple BoundArchs per triple as a future development. 
This fixes an issue where spirv and nvptx swap BoundArch's, causing failure cases. #3631
This patch should also resolve #4662

Some tests were changed in sycl-offload-with-split.c where the output from the device compilation phase was being used as an input by the wrong device. 3 tests are also introduced, two check the ordering behaviour of the triples and the bounded arch. The third checks for behaviour when targetting spirv, nvptx, and amdgcn.