[SYCL] Detect changes in sycl headers
When subdirectory of the DEPENDS directories are modified the change is not
detected by CMake. The change set full list of files in subdirectory to
DEPENDS to detect the change. file(GLOBE_RECURSE) is used with
CONFIGURE_DEPENDS option to rerun configure if new entry is added to
directory structure.
Restore version.hpp in source directory as it is used to build SYCL
library.

Removing header file is the only case when change is not propagated to
build and install directory. This problem existed before the change and
TODO was left to fix it in future.