[SYCL] Prevent fallback assert postprocessor being passed to CUDA and HIP devices (#4604)
This patch fixes an error with HIP backend where tests DeviceCodeSplit/split-per-kernel.cpp and DeviceCodeSplit/split-per-source-main.cpp fail with `Memory access fault by GPU node-4 (Agent handle: 0x709c50) on address (nil). Reason: Page not present or supervisor privilege.` 

This is due to a postprocessor being sent to `queue_impl::submit_impl`. 
The patch add checks to `queue.hpp` so postprocessors needed for fallback assert do not get passed for HIP and CUDA devices. Neither currently support fallback asserts.

Note: adding `__AMDGCN__` to
```
#if !defined(SYCL_DISABLE_FALLBACK_ASSERT) && !defined(__NVPTX__)
#define __SYCL_USE_FALLBACK_ASSERT 1
#else
#define __SYCL_USE_FALLBACK_ASSERT 0
#endif
```
was not a fix as a new error arose in its place. Investigation showed that macros `__AMDGCN__` and `__NVPTX__` are only set for device compilation and as a result did not stop preprocessors being passed to `submit_impl`. I think the postprocesor was not supposed to be passed to CUDA devices. Despite being passed it was not causing an error for CUDA just HIP.