[SYCL] Detect changes in sycl headers (#4723)
This is follow up for #4710

When a sub-directory of the DEPENDS directories is modified the change is not
detected by CMake. The change adds a full list of files in the sub-directory to
DEPENDS to detect the change. file(GLOBE_RECURSE) is used with
CONFIGURE_DEPENDS option to rerun configure step if a new entry is added to
the directory structure.
Restore version.hpp in source directory as it is used to build SYCL
library.

Removing a header file is the only case when change is not propagated to
the build and install directory. This problem existed before the change and
TODO was left to fix it in the future.