[SYCL] Fix calculation of unreleased events in a pool
Before the change number of unreleased events was set to the maximum at
event pool creation and then only decremented at the moment when event
from the pool is released. This is not correct for the case when pool
is not completely full.
Patch fixes calculation of unreleased events in a pool.

Signed-off-by: Artur Gainullin <artur.gainullin@intel.com>