[SYCL] Enable diagnosing of indirect implicit capture of "this"
The compiler issues an error if a lambda passed directly to a SYCL kernel
function implicitly captures "this", but does not issue an error if the
lambda that captures "this" is itself captured by the kernel lambda.
The situation when SYCL kernel function captures lambda is a normal
workflow for SYCL reductions, so it is important to diagnose this case
as well.