[SYCL] Use device code cache when building kernel bundles (#4724)
The persistent device code cache, when enabled, is not used when
building the underlying native program in a kernel bundle. These changes
unify cache lookup with a fallback to program creation from
ProgramManager::getBuiltPIProgram with the program creation in
ProgramManager::build, ensuring that the latter also uses cached device
code. Likewise, ProgramManager::build will write the device code to the
persistent cache if it was not there already.

Signed-off-by: Steffen Larsen <steffen.larsen@intel.com>