[SYCL] Enforces exception on empty device vector in sycl::link (#4760)
The SYCL 2020 specification states that sycl::link must fail with a SYCL
exception with errc::invalid if the vector of devices is empty. However
the implementation of sycl::link skips this check if the list of kernel
bundle objects is empty as well. These changes move the aforementioned
check for devices earlier in the implementation of sycl::link and throws
an exception even if there are no kernel bundle objects specified.

Signed-off-by: Steffen Larsen <steffen.larsen@intel.com>