[SYCL] Rename cuda, hip and fix renaming for level_zero backends
According to SYCL spec, backend::cuda is renamed to backend::ext_oneapi_cuda,
backend::hip is renamed to backend::ext_oneapi_hip. Old namings were marked as
deprecated. The renaming is also applied in the sources, tests and unit-tests.
Renaming for the level_zero backend is also applied for the sources, tests
and unit-tests.