Update SPV and OCL specs w.r.t. variable name
We decided that we need some alternate way for the host to reference a
global variable in the kernel other than using the exported SPIR-V
name.  Extend the `HostAccessINTEL` decoration to take a new `Name`
parameter.  This allows the host to access even a non-exported
`OpVariable`.

Changes are still needed to the "DeviceGlobal.md" spec to explain how
this new field will be set by the compiler.  Those changes will be made
in a separate commit.