[Driver][SYCL] -Xsycl option triple check incorrect with multiple -fsycl-targets
When passing multiple -fsycl-targets on the command line, the last one wins.
There is a check for the number of target triples provided by the user when
applying -Xsycl-target* options to the device compilation.  We were
improperly counting all of the -fsycl-targets values instead of only taking
into account the last one only.