Remove duplicates for feature test macros
There were duplicates for the following two feature test macros:
 - SYCL_EXT_ONEAPI_DEFAULT_CONTEXT (in CL/sycl/platfrom.hpp)
 - SYCL_EXT_ONEAPI_GROUP_SORT (in sycl/ext/oneapi/group_sort.hpp)

The file 'feature_test.hpp.in' is the single point to define all the
feature test macros. The user should have ability to test a feature
test macro before to include the corresponding header file; therefore,
the macro must be defined outside of the header that implements the
feature under question.