[SYCL] Include moved cuda backend header in feature_test.hpp.in
When the old cuda backend header (CL/sycl/backend/cuda.hpp) is included,
a deprecation warning is generated. When "to trait warning as errors"
compilation flag is enabled, the warning leads to a compilation error.
The new path to the cuda backend header (sycl/ext/oneapi/backend/cuda.hpp)
must be used to avoid the error.