[SYCL][DOC] Update CUDA docs with Windows support details (#4796)
This patch updates the compiler and runtime docs and getting started docs with details of CUDA's support for Windows.

It explains the motivation and usage of re-mangled `libspirv-nvptx64--nvidiacl.bc` variants. 
Links to url for installing CUDA on Windows OS.
Clearly states that there should be no dependencies on Linux system.