[Driver][SYCL] Fix -Xsycl option triple check with multiple -fsycl-targets (#4789)
When passing multiple -fsycl-targets on the command line, the last one wins.
There is a check for the number of target triples provided by the user when
applying -Xsycl-target* options to the device compilation.  We were
improperly counting all of the -fsycl-targets values instead of only taking
into account the last one only.