[Driver][SYCL][FPGA] Implied default device forces emulation (#4801)
When using -Xshardware or -Xssimulation on the command line with
FPGA, the device AOT compilation should be performed by aoc.  This
was not occuring due to the implied default device when one such
object or archive was supplied on the command line.
Make adjustments to our discovery of FPGA hardware/simulation mode
by only checking the actual FPGA toolchain and not just the first
offload toolchain encountered.