[sycl-post-link] Refactor cloned modules processing
ValueToValueMapTy map is not fully release on destruction the data that is
allocated on CloneModule call. Also split modules may utilize tens of megabytes
and their amount may be significant. It all may lead to overflow of RAM memory.

Rewrite module splitting in a way that each split module is processed and
immediately saved into a file without temporary collecting all modules in RAM.

Signed-off-by: Mikhail Lychkov <mikhail.lychkov@intel.com>