[SYCL] Create static shared-release handler after loading plugins
To enforce default contexts to be released before global
variables in plugins and associated libraries, these changes intializes
a function-local static variable after the plugins have been loaded.
Due to the guarantees of std::exit the destructor of the new
function-local variable will be called before the destructor of all
global variables initialized prior to the function-local static
variable's initialization.

Signed-off-by: Steffen Larsen <steffen.larsen@intel.com>