[SYCL] Remove duplicate devices on submission to kernel_bundle API functions (#4790)
When submitting devices to compile, section 4.11.7 of the SYCL 2020 spec states that the compilation is for each unique device, duplicate devices removed. Here we ensure that the devices sent to the backend are each unique.
Also, Level Zero doesn't support compilation for more than one device but it calls die() presently. Changing that as well.

Signed-off-by: Chris Perkins <chris.perkins@intel.com>