[CI] Add reusable workflow for SYCL builds (#4773)
The purpose of this new reusable workflow is to be a "workflow library" to allow calling various build configurations from other workflows. Current limitation of reusable workflows is that they can not be called from "inside" of pull request. To workaround it, this workflow is run when it is changed in pull request.

See official documentation for more info: https://docs.github.com/en/actions/learn-github-actions/reusing-workflows

The next pull request will switch post-commit job to use this library, as well as add Docker container build job with pre-built SYCL.