[sycl-post-link] Add padding to default values of structs
Some structs may require additional padding at the end of their default
values to correctly fill the size they are reported to have. These
changes enforce additional padding on default values for structs when
the added bytes do not correspond to the reported size of the struct.
This ensures correct offset of values, corresponding to those expected
by both the DPC++ runtime and the generated SYCL kernels.

Signed-off-by: Steffen Larsen <steffen.larsen@intel.com>