[SYCL] Workaround: mark sycl::vec as device copyable on host
See commit https://github.com/intel/llvm/commit/43c6d7a4e1f78, the
sycl::vec class is not trivially copyable on the host side, so it is
not considered as device copyable. A partial specialization for the
sycl::is_device_copyable for sycl::vec is implemented to mark the class
as device copyable on the host side.