[SYCL] Fix hexadecimal byte streaming in device binary properties
Using the streaming operator with DeviceBinaryProperty that contains
a byte array will try to write the bytes as hexadecimal values to the
associated stream. However, std::hex only converts integer values whilst
the bytes the operator writes are chars. These changes cast the chars to
integer values to ensure that the values are correctly written with
hexadecimal representation.

Signed-off-by: Steffen Larsen <steffen.larsen@intel.com>