[SYCL][HIP] Disable partial global offset support
This patch is disabling global offset support for the HIP plugin.

The existing `libclc` implementation for `__spirv_GlobalOffset_x()` was
using the HIP implicit args system, but these are not currently setup by
`clang` nor by the HIP plugin when running with SYCL.

Additionally the global offset is used when computing the global id,
which means that even in kernels that don't use global offsets the
kernel would still read a global offset from the uninitialized implicit
args, which in some cases would cause crashes.

Additionaly the HIP plugin is trying to mimick the CUDA plugin
behaviour, but the PTX backend has an extra IR pass to generate a
specific wrapper for kernels using global offsets, which is not part of
the AMDGCN compilation pipeline so when using a global offset different
from 0, the HIP plugin would just try to call a non-existant kernel.

So this patch is adding an assert in the HIP plugin when trying to use
global offsets different from 0, and forces the global offsets to be 0
in the kernel, until we implement proper support for global offsets.