[SYCL] Make a test class trivially copyable in multi_ptr.cpp (intel/llvm-test-suite#555)
The class 'point' is used to test 'multi_ptr'. A 'sycl::buffer' for instances
of the class is created and the class must be a trivially copyable to be trait
as device copyable. Otherwise a compilation error will be generated after
a new check for device copyability in the 'sycl::buffer' class, see
intel/llvm#4914.

Signed-off-by: Pavel Samolysov <pavel.samolysov@intel.com>