[SYCL] Diagnose non-forward declarable kernel name types
Add additional error emission in host mode
because only on host with help of integration header possible to
differentiate case like this:
```
int main() {
  parallel_for<class KernelName>(..);
}
```
which uses technically non-forward declarable kernel name type but
still allowed by SYCL spec, from case like this:
```
int main() {
  class KernelName;
  parallel_for<class KernelName>(..);
}
```
which should be diagnosed, since the errors are emitted in runtime.

Everything works when `KernelName` typename is used directly in
`parallel_for`, because in this case when the type `KernelName` is
forward declared by integration header, host uses `::KernelName`
typename to access kernel info.
However when `KernelName` is forward declared in non-global/namespace
scope it actually produces a separate delclaration `main::KernelName`
which is not visible for runtime code that submits kernels.