[SYCL][Doc] Deprecate SubGroup extension (#4943)
All functionality introduced by this extension that is currently
implemented in DPC++ has equivalent functionality in SYCL 2020:
- The sub_group class
- The group functions (e.g. broadcast, barrier)
- The group algorithms (e.g. permutes, shifts, reductions)

All functionality related to "primary" sub-group sizes is not
implemented. Rather than maintain a list of which features are
extensions and which are implemented, now seems a good chance to
deprecate the whole extension. The "primary" sub-group size
functionality will be reintroduced as an extension against SYCL 2020,
aligned with more recent developments (e.g. using compile-time
properties instead of new attributes).

Signed-off-by: John Pennycook <john.pennycook@intel.com>