[SYCL] Check if the underlying buffer type is device copyable (#4914)
The implementation should diagnose an error if the underlying type "T"
of a buffer is not device copyable. The spec says that this is
a requirement for all buffers. It does not make any exception for
buffers that are used only on the host:

"The underlying data type of a buffer T must be device copyable as
defined in Section 3.13.1."

(https://www.khronos.org/registry/SYCL/specs/sycl-2020/html/sycl-2020.html#subsec:buffers)