[SYCL][Doc] Clarify restrictions on device global variables (#4697)
Clarify the restrictions for declaring variables of type
`device_global`.  These restrictions are similar to those we have
already for `specialization_id` variables.

The restriction against arrays of `device_global` variables is not
strictly necessary, but it would take extra effort to implement and
we do not have a compelling use case.  We could lift the restriction
later if we find a need.