Revert "[SYCL] Check if the underlying buffer type is device copyable (#4914)"
Since sycl::vec is not a trivially copyable class, even though we
have defined a specialization for sycl::is_device_copyable template
for the sycl::vec class itself, it doesn't make any classes that have
members of the sycl::vec type device copyable since those classes aren't
trivially copyable yet. For example, a class from SYCL-CTS:

template <typename T>
struct image_accessor_failure_item {
  bool triggered;
  T value;
  T expected;

  image_accessor_failure_item() :
    triggered(false),
    value(0),
    expected(0) {}
};

is not trivially copyable since it has two fields: value and expected of
type sycl::vec; therefore, the class is not device copyable and a
compilation error occurs when a sycl::buffer is created for this class
instances.

The regression makes us to revert the commit and delay adding the check
until the sycl::vec class is modified to be trivially copyable.

This reverts commit 61f1ae6a9ef87cbed8c14713995e9292557d3b17.