[SYCL] Check if the underlying buffer type is std::string
The implementation should diagnose an error if the underlying type "T"
of a buffer is not device copyable. Currently, we cannot implement a
check for all not device copyable types since sycl::vec is not
trivially copyable on the host and therefore all the types that contain
sycl::vec are also not trivially copyable as well as not device
copyable. The check is implemented for std::string only.