[sycl-post-link] Fix memory growth on module split
Module splitting is performed via llvm::CloneModule() call. When cloning module
MDNode* instances for split module are allocated in source module's LLVMContext
and kept in memory after split module is destroyed. In case if there are
a lot of split modules and/or they have a lot of metadata nodes it may lead to
memory overflow.
Add optional mode which periodically creates new context and split modules in
that context. It fixes memory growth, but at the cost of greatly increased
processing time.
This mode is off by default and can be turned on via option
'-reduce-memory-usage=true'.

Signed-off-by: Mikhail Lychkov <mikhail.lychkov@intel.com>