[LIBCLC] Make sure flags are unset when building libclc for AMD (#4991)
Without this patch the flags always contain `-mcpu=tahiti`. The problem occurs when llvm compiles the spirv module it chooses `llvm::AMDGPUSubtarget::SOUTHERN_ISLANDS` as the subtarget. The instruction encoding differers, and when targeting `llvm::AMDGPUSubtarget::GFX9` the final binary is invalid. This in turns causes the `HSA_STATUS_ERROR_ILLEGAL_INSTRUCTION`.