[WIP Do not merge!][SYCL][MATRIX][CUDA]
Initial implementation of bf16 data format using a new interface that make use of the matrix_type enum.
matrix_type enum is used as the first template parameter instead of typename in e.g. joint-matrix-load to allow the user to provide fp32 matrices that the implementation is responsible for converting to special format types bf16 or tf32 prior to a matrix multiplication operation that returns an fp32 matrix.
The conversion to the more efficient types will be done if the user specifies matrix_type::bf16.
This means that the user does not have to worry about conversion from fp32 to bf16 when there is no advantage for to the user to do this themselves.

Signed-off-by: jack.kirk <jack.kirk@codeplay.com>