[LIBCLC] Fix config.h include
This is fixing building libclc for `amdgcn-amdhsa`.

It showed up in the `ldexp.cl` builtin.

The issue is that `_CLC_DECL` is declared in `func.h` and in other files
`config.h` relies on other `func.h` being included first. This is
confusing and error prone so include `func.h` directly in `config.h`
instead.